Remove "--terse", but add `--stress` mode that will discard output.
