+++ pkg/bench/benchmark.go
@@ -47,7 +47,6 @@ type Benchmark interface {
 
 // Common contains common benchmark parameters.
 type Common struct {
-
 	// Default Put options.
 	PutOpts minio.PutObjectOptions
 
@@ -64,6 +63,8 @@ type Common struct {
 	Error func(data ...interface{})
 
 	Client func() (cl *minio.Client, done func())
+
+	Collector *Collector
 
 	Location string
 	Bucket   string
@@ -86,8 +87,8 @@ type Common struct {
 	// Clear bucket before benchmark
 	Clear bool
 
-	// Terse output.
-	Terse bool // indicates if we prefer a terse output useful in lengthy runs
+	// DiscardOutput output.
+	DiscardOutput bool // indicates if we prefer a terse output useful in lengthy runs
 
 	// Does destination support versioning?
 	Versioned bool
@@ -231,5 +232,13 @@ func (c *Common) prepareProgress(progress float64) {
 	select {
 	case c.PrepareProgress <- progress:
 	default:
+	}
+}
+
+func (c *Common) addCollector() {
+	if c.DiscardOutput {
+		c.Collector = NewNullCollector()
+	} else {
+		c.Collector = NewCollector()
 	}
 }
