+++ pkg/bench/mixed.go
@@ -162,7 +162,7 @@ func (g *Mixed) Prepare(ctx context.Context) error {
 	console.Info("\rUploading ", g.CreateObjects, " objects of ", src.String())
 	var wg sync.WaitGroup
 	wg.Add(g.Concurrency)
-	g.Collector = NewCollector()
+	g.addCollector()
 	obj := make(chan struct{}, g.CreateObjects)
 	for i := 0; i < g.CreateObjects; i++ {
 		obj <- struct{}{}
@@ -262,9 +262,6 @@ func (g *Mixed) Start(ctx context.Context, wait chan struct{}) (Operations, erro
 						ObjPerOp: 1,
 						Endpoint: client.EndpointURL().String(),
 					}
-					if g.Terse {
-						op.File = ""
-					}
 
 					op.Start = time.Now()
 					var err error
@@ -339,9 +336,6 @@ func (g *Mixed) Start(ctx context.Context, wait chan struct{}) (Operations, erro
 						File:     obj.Name,
 						ObjPerOp: 1,
 						Endpoint: client.EndpointURL().String(),
-					}
-					if g.Terse {
-						op.File = ""
 					}
 
 					op.Start = time.Now()
