+++ pkg/bench/retention.go
@@ -33,8 +33,7 @@ import (
 // Retention benchmarks download speed.
 type Retention struct {
 	Common
-	Collector *Collector
-	objects   generator.Objects
+	objects generator.Objects
 
 	CreateObjects int
 	Versions      int
@@ -61,7 +60,7 @@ func (g *Retention) Prepare(ctx context.Context) error {
 	console.Info("\rUploading ", g.CreateObjects, " objects with ", g.Versions, " versions each of ", src.String())
 	var wg sync.WaitGroup
 	wg.Add(g.Concurrency)
-	g.Collector = NewCollector()
+	g.addCollector()
 	obj := make(chan struct{}, g.CreateObjects)
 	for i := 0; i < g.CreateObjects; i++ {
 		obj <- struct{}{}
@@ -98,9 +97,6 @@ func (g *Retention) Prepare(ctx context.Context) error {
 						File:     obj.Name,
 						ObjPerOp: 1,
 						Endpoint: client.EndpointURL().String(),
-					}
-					if g.Terse {
-						op.File = ""
 					}
 
 					opts.ContentType = obj.ContentType
